==> code/notebooks/00_kjsailer_project.ipynb <==
# Project: Levels of Football

Kaleb Sailer

This is the scraping file for high school data.

### Packages

import pandas as pd 
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
import time
import random

## High School Data: Bound Website

## using 2023-2024 data so it does not change with the playoffs currently going on

# team name and record scraping 
url = 'https://www.gobound.com/ia/ihsaa/football/2023-24/powerindex'

service = Service(ChromeDriverManager().install())
browser = webdriver.Chrome(service=service)

team_names = []
wins = []
losses = []
ppg = []
ppg_a = []
ppg_diff = []

browser.get(url) 

# waiting for page
time.sleep(2)

# scrape the team name
team_elements = browser.find_elements(By.XPATH, "//table//tr/td[2]/a")

for team_name in team_elements:
        team_names.append(team_name.text)

# scrape the wins
win_elements = browser.find_elements(By.XPATH, "//table//tr/td[5]")

for win_count in win_elements:
        wins.append(win_count.text)

# scrape the losses
loss_elements = browser.find_elements(By.XPATH, "//table//tr/td[6]")

for loss_count in loss_elements:
        losses.append(loss_count.text)

# scrape the ppg
ppg_elements = browser.find_elements(By.XPATH, "//table//tr/td[39]")

for ppg_count in ppg_elements:
        ppg.append(ppg_count.text)

# scrape the ppg allowed
ppg_a_elements = browser.find_elements(By.XPATH, "//table//tr/td[40]")

for ppg_a_count in ppg_a_elements:
        ppg_a.append(ppg_a_count.text)

# scrape the ppg diff
ppg_diff_elements = browser.find_elements(By.XPATH, "//table//tr/td[41]")

for ppg_diff_count in ppg_diff_elements:
        ppg_diff.append(ppg_diff_count.text.strip('+'))
    
browser.quit()

# data frame creation
hs_team_stats_df = pd.DataFrame({
    'team_name': team_names,
    'W': wins,
    'L': losses,
    'ppg': ppg,
    'ppg_a': ppg_a,
    'ppg_diff': ppg_diff
})

display(hs_team_stats_df.head())

id_metrics = ['h20200131032829975bb9dfada0d8e49', 'h2020013103313188014d330346cea4b', 'h20200131080614042e9114503743541']

service = Service(ChromeDriverManager().install())
browser = webdriver.Chrome(service=service)

team_data = {}

for id_metric in id_metrics:
    url = f'https://www.gobound.com/ia/ihsaa/football/2023-24/leaderlist?competitor=team&range=season&idGroup=h202302090257205525ac5f791e53a41&idMetricGroup={id_metric}&page=1&block=total'
    browser.get(url)

    time.sleep(5)
    
    rows = browser.find_elements(By.XPATH, '/html/body/div[2]/div/div[2]/div/div[1]/div[2]/div/div/div/div[2]/table/tbody/tr')
    print(f"Found {len(rows)} rows for metric {id_metric}")

    for row in rows:
        team_element = row.find_element(By.XPATH, './td[2]/div/div/h7/a')
        team_name = team_element.text
        print(f"Processing team: {team_name}")

        if team_name not in team_data:
            team_data[team_name] = {}

        if id_metric == 'h20200131032829975bb9dfada0d8e49':
            pass_tds = row.find_element(By.XPATH, './td[8]').text
            pass_yds = row.find_element(By.XPATH, './td[6]').text
            team_data[team_name]['pass_tds'] = pass_tds
            team_data[team_name]['pass_yds'] = pass_yds

        elif id_metric == 'h2020013103313188014d330346cea4b':
            rush_tds = row.find_element(By.XPATH, './td[6]').text
            rush_yds = row.find_element(By.XPATH, './td[4]').text
            team_data[team_name]['rush_tds'] = rush_tds
            team_data[team_name]['rush_yds'] = rush_yds

        elif id_metric == 'h20200131080614042e9114503743541':
            off_tds = row.find_element(By.XPATH, './td[4]').text
            team_data[team_name]['off_tds'] = off_tds
        
        time.sleep(random.uniform(1, 3))
        
browser.quit()

team_metric_df = pd.DataFrame.from_dict(team_data, orient='index')
team_metric_df.head(1)

### Concatenation of the two High School Data Frames

hs_stats_df = pd.merge(hs_team_stats_df, team_metric_df, left_on='team_name', right_on=team_metric_df.index, how='inner')
hs_stats_df = hs_stats_df[['team_name', 'W', 'L', 'ppg', 'ppg_a', 'ppg_diff', 'pass_yds', 'pass_tds', 'rush_yds', 'rush_tds', 'off_tds']]

# type conversions
hs_stats_df[['ppg', 'ppg_a', 'ppg_diff']] = hs_stats_df[['ppg', 'ppg_a', 'ppg_diff']].astype('float')

hs_stats_df[['W', 'L', 'pass_yds', 'pass_tds', 'rush_yds', 'rush_tds', 'off_tds']] = hs_stats_df[['W', 'L', 'pass_yds', 'pass_tds', 'rush_yds', 'rush_tds', 'off_tds']].astype('int')

hs_stats_df.to_csv('hs_stats_raw.csv', encoding='utf-8', index=False)
